backend/pyproject.toml
[tool.poetry]
name = "talent-pool-backend"
version = "0.1.0"
description = "Talent Pool backend with Django + DRF"
authors = ["Arif <arifthecoder@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
djangorestframework = "^3.14"
psycopg2-binary = "^2.9"
django-cors-headers = "^4.3"
python-dotenv = "^1.0"
gunicorn = "^21.2"
djangorestframework-simplejwt = "^5.3.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-django = "^4.7"
black = "^24.0"
isort = "^5.13"
